remove keycast dkg
